The Final Version of My Prime Factorization
For Windows, download and run PrimeFactorization 1.3.3.exe

For any other operating system, download PrimeFactorization(1.3.3).jar
Run it by doing Java -jar PrimeFactorization(1.3.3).jar in a terminal or
command promt.
If you want to run it on windows but not use the .exe download both run
PrimeFactorization(1.3.3).bat and PrimeFactorization(1.3.3).jar After
that put them in the same directory and run the .bat file.